+++ gitlias/config.go
@@ -1,4 +1,4 @@
-package main
+package gitlias
 
 import (
 	"os"
