+++ gitlias/gitlias_test.go
@@ -1,0 +1,3 @@
+package gitlias
+
+// TODO: Testing
